- r8574 fix: IMP creation with HP < OKLIFE possible (by Flugente) - r8020 fix: possible overflow when creating building (by asm-debacle) (by Flugente) - r8026 fix: It is possible to leave a sector while an explosion has been activated but not yet occurred. In this case the game waits for the explosion to occur before handling AI, resulting in a deadlock. (by Flugente)

+++ TileEngine/Buildings.cpp
@@ -553,8 +553,8 @@ BUILDING * GenerateBuilding( INT32 sDesiredSpot )
 	// something that has gone pear shaped
 	if(uiLoop2 >= WORLD_MAX)
 	{
-		UINT8 x = 0;
-		UINT8 y = 0;
+		INT32 x = 0;
+		INT32 y = 0;
 		while((sDesiredSpot - ((y + 1) * WORLD_COLS)) >= 0)
 		{
 			y++;
